requirements.txt
Flask
requests
PyJWT
pymongo
werkzeug
gunicorn
fastapi
uvicorn[standard]
python-dotenv
requests
bcrypt